app/backend/pyproject.toml
[tool.poetry]
name = "claims-backend"
version = "0.1.0"
description = "Claim Resubmission Ingestion Pipeline - Backend"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
python-dateutil = "^2.9.0.post0"
prometheus-client = "^0.20.0"
structlog = "^24.4.0"
orjson = "^3.10.7"
ijson = "^3.2.3"
rapidfuzz = "^3.9.6"
pandas = "^2.2.2"
polars = "^1.5.0"
prefect = "^3.0.0"
python-multipart = "^0.0.9"
httpx = "^0.27.2"
pymongo = "^4.8.0"
dnspython = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
hypothesis = "^6.112.0"
mypy = "^1.11.1"
ruff = "^0.6.1"
black = "^24.8.0"
types-python-dateutil = "^2.9.0.20240821"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E203", "E266", "E501", "W503"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true
disallow_untyped_defs = true
no_implicit_optional = true


